Require instantiation modules be inline and empty
They must be inline to avoid the ambiguity when there exists
a real module on the same file path. Also, usage of non-inline
modules is not possible in current stable Rust.

They must be empty, otherwise there could be definitions or
use statements that silently mask the glob imports from the
macro invocation's root module.