Forbid using instantiate_tests as inner attribute
The type paths referred in the attribute should belong to the
parent module. Usage of inner attribute syntax would obscure
the provenance of the paths, and we can't have the instantiation modules
be populated with their own use declarations, which anyway
would have to come after the inner attributes, also obscuring
the matter.