Add a disabled test for expected compile failure
When a name defined in the parent of an instantiation module
shadows a name in the root module of the macro invocation,
its use in the instantiation arguments causes a compilation error
due to the name being ambiguous between the two glob imports.

Add a test case hidden by the  "test-compile-fail" feature gate
to exercise the failure.