Enforce uniform number of generic parameters
Emit an error when a test function in the module has a different
number of generic arguments than other previously processed functions.
This is a quality of life improvement: the test writer is pointed at a
disparity in a specific function, rather than being told that some
function call in the instantiation module can't be qualified with
the supplied type arguments.