Deconflict glob imports in deeply nested modules
Solve the problem of having to use names from both the macro invocation
root module where the generic tests are defined, and the possibly
nested parent module of an instantiation module.

Do this by generating a shim module inside each instantiated test
function, where the instantiation parent module's names are imported
to be used in the type-qualified call to the generic test function.
The types of function's arguments and return value are not directly
referred to in the shim, instead these values are marshalled into
types generated for each call signature that the macro has seen on
the test functions, so that fully qualified names of these
signature types can be used instead. The instantiated test function,
for its part, uses unqualified names imported from the root and
calls the shim with a populated argument signature struct.

Reorg extraction code into a separate mod extract.
The expand module is now more concentrated on code generation.