Comment about non-importance of generic arg order
Use of HashSet means the order is random and varies with
each invocation. This has hardly any significance as it should be
rare to have more than one lifetime, the generated shims are
not seen anywhere except backtraces and call stacks,
and nobody should care about such such a minor detail.