Special case error message for wildcard arguments
We can't support wildcard arguments since the generic function
needs to be called with something, but replacing the wildcard
with a named variable binding changes semantics: the bound value
is not dropped until the return from the instantiated function.

At least describe this pattern as specifically forbidden in the
error message.