Rework error reporting
There is no need to validate ahead of extracting attributes
from the AST, as the span problems are not related to the macro's
tree modifications.
What does makes sense is to continue processing functions regardless
of errors in preceding functions, and output the combined errors.
Also, don't needlessly calculate the span for the whole invocation
module when the success case does not need it.